CDAP-15836. Move packages and pom changes
