+++ src/main/java/io/cdap/plugin/date/DateTransform.java
@@ -14,8 +14,10 @@
  * the License.
  */
 
-package io.cdap.plugin;
+package io.cdap.plugin.date;
 
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Strings;
 import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Macro;
 import io.cdap.cdap.api.annotation.Name;
@@ -28,8 +30,6 @@ import io.cdap.cdap.etl.api.InvalidEntry;
 import io.cdap.cdap.etl.api.PipelineConfigurer;
 import io.cdap.cdap.etl.api.Transform;
 import io.cdap.cdap.etl.api.TransformContext;
-import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Strings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
